use int64_t in prototypes

+++ lib/upipe-hbrmt/upipe_sdi_dec.c
@@ -152,7 +152,7 @@ struct upipe_sdi_dec {
     const char *output_chroma_map[UPIPE_SDI_DEC_MAX_PLANES];
 
     /** UYVY to V210 */
-    void (*uyvy_to_v210)(const uint16_t *y, uint8_t *dst, ptrdiff_t width);
+    void (*uyvy_to_v210)(const uint16_t *y, uint8_t *dst, int64_t width);
 
     /** UYVY to 8-bit Planar */
     void (*uyvy_to_planar_8)(uint8_t *y, uint8_t *u, uint8_t *v, const uint16_t *l, const int64_t width);
@@ -1075,7 +1075,7 @@ static void uyvy_to_planar_10_c(uint16_t *y, uint16_t *u, uint16_t *v, const uin
         dst += 4;                       \
     } while (0)
 
-static void uyvy_to_v210_c(const uint16_t *y, uint8_t *dst, ptrdiff_t width)
+static void uyvy_to_v210_c(const uint16_t *y, uint8_t *dst, int64_t width)
 {
     uint32_t val, tmp1, tmp2;
     int i;
